Give the screenshot skill an ad-hoc lane and a truthful label
Three corrections, each from hitting the limit in practice.

Publishing said to clone the repo and push, which an agent is often not
permitted to do outside the project; it now writes the images with the
git data API, and checks a raw URL before posting a comment full of
possibly-broken images.

A surface outside the catalogue could only be captured by extending the
catalogue permanently. Extract the capture helpers so a throwaway spec
can import them, and document the one-off lane beside the permanent one
- including that the two specs must never run in one command, since the
catalogue wipes the shared output directory and the config is
fullyParallel.

The label read as a record of screenshots attached when it was a request
for them. Ask with `screenshots needed` and swap it for `screenshots`
once a comment with images actually goes up.

Also note what the harness cannot see: macOS hides scrollbars unless a
mouse is attached, and a styled ::-webkit-scrollbar does not override
it, so any gutter or track change captures identically on both sides.
